requirements.txt
openai==1.100.2
python-dotenv==1.1.1
pandas==2.3.1
reportlab==4.4.3
